feat: allow signing/encrypting Buffer payloads
The reference PHP implementation[1] allows "to store arbitrary binary
data in a PASETO, by invoking the Protocol classes directly". This
change adds support for the V1.{sign,encrypt}/V2.sign functions to be
able to receive Buffers as the payload. It also adds a new option (`{
buffer: true }`) to the V1.{verify,decrypt}/V2.verify functions so that
they return a Buffer instead of a parsed token, for symmetry.

1: https://github.com/paragonie/paseto/tree/master/docs/02-PHP-Library#using-the-protocol-directly